Move config to src folder
